+++ src/services/ElementSources.php
@@ -429,16 +429,23 @@ class ElementSources extends Component
      * @param class-string<ElementInterface> $elementType The element type class
      * @param string $sourceKey The element type source key
      * @param string[]|null $customAttributes Custom attributes to show rather than the defaults
+     * @param FieldLayout[]|null $fieldLayouts The field layouts that should be factored in
      * @return array[]
      */
-    public function getTableAttributes(string $elementType, string $sourceKey, ?array $customAttributes = null): array
-    {
+    public function getTableAttributes(
+        string $elementType,
+        string $sourceKey,
+        ?array $customAttributes = null,
+        ?array $fieldLayouts = null,
+    ): array {
         // If this is a source path, use the first segment
         if (($slash = strpos($sourceKey, '/')) !== false) {
             $sourceKey = substr($sourceKey, 0, $slash);
         }
 
-        if ($sourceKey === '__IMP__') {
+        if ($fieldLayouts !== null) {
+            $sourceAttributes = $this->getTableAttributesForFieldLayouts($fieldLayouts);
+        } elseif ($sourceKey === '__IMP__') {
             $sourceAttributes = $this->getTableAttributesForFieldLayouts($elementType::fieldLayouts(null));
         } else {
             $sourceAttributes = $this->getSourceTableAttributes($elementType, $sourceKey);